requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow==2.13.0
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psutil==5.9.6
pytest==7.4.3
httpx==0.25.2